requirements.txt
PySide6>=6.6
requests>=2.31
python-vlc>=3.0.0